build: Expose getMakerHash function
