+++ src/__tests__/eip712.test.ts
@@ -2,8 +2,8 @@ import { expect } from "chai";
 import { utils } from "ethers";
 import { TypedDataDomain } from "@ethersproject/abstract-signer";
 import { setUpContracts, SetupMocks, getSigners, Signers } from "./helpers/setup";
-import { computeDigestMaker, getDomainSeparator, getMakerHash } from "./helpers/eip712";
-import { contractName, version } from "../constants/eip712";
+import { computeDigestMaker, getDomainSeparator } from "./helpers/eip712";
+import { contractName, version, getMakerHash } from "../constants/eip712";
 import { SupportedChainId, Maker, CollectionType, QuoteType } from "../types";
 
 describe("EIP-712", () => {
