+++ src/index.ts
@@ -1,6 +1,5 @@
 export { addressesByNetwork } from "./constants/addresses";
 export { chainInfo } from "./constants/chains";
-export * as eip712 from "./constants/eip712";
 
 import * as tokens from "./utils/calls/tokens";
 import * as exchange from "./utils/calls/exchange";
@@ -10,15 +9,19 @@ import * as orderValidator from "./utils/calls/orderValidator";
 import * as strategies from "./utils/calls/strategies";
 import * as encode from "./utils/encodeOrderParams";
 import * as signMakerOrders from "./utils/signMakerOrders";
+import * as eip712 from "./constants/eip712";
 const utils = {
-  ...tokens,
+  calls: {
+    ...tokens,
+    ...exchange,
+    ...nonces,
+    ...transferManager,
+    ...orderValidator,
+    ...strategies,
+  },
+  eip712,
   ...encode,
-  ...exchange,
-  ...nonces,
-  ...transferManager,
-  ...orderValidator,
   ...signMakerOrders,
-  ...strategies,
 };
 export { utils };
